requirements.txt
easyocr
Pillow
numpy
roboflow
streamlit